THIS IS A REALLY BIG UPDATE.
the biggest changes are probably the player queue system and all the
changes to player[].status that were required. lots of thread
synchronization was also added. a new module (playerdata) was created to
hold the player array as well as locking functions. fixes to the arena
queue system too. various other bugfixes.

it compiles now, but logins don't work yet. bugfixing to come.