added interface priorities, in the interface header, and also moved the interface id into the header.
added auth_prefix, implementing mikethenose's request: the characters "+->@#$"
are now reserved prefix characters. if you have one of them in front of your
name, it will be send to the biller for authentication without the character,
and if you have permission (prefix_+), it will be added back again.

added new field to AuthData, sendname, separate from name. this is to handle
the cases (like auth_prefix) when you want other players to see someone with
one name, but handle him internally (for group assignment, persistant data,
etc.) with another.

added HasCapabilityByName to capman, for times when you want to check
capabilities for a player who's not logged in.

updates to TODO

skeleton for devguide.tex file