cleanups in Pppk.
added option of sending a percentage of position packets with antiwarp
to the whole arena. this is to fix the issue where someone warps to a
port in an antied arena and quickly lays another before he knows the
area is antied.

added ?help and added help text to most commands.

added extract-docs.py.

moved userguide.tex into its own directory, with a makefile, and made it
use the output of extract-docs.py for the command section.