in game.c, don't ever send a packet with both NET_RELIABLE and NET_DROPPABLE set. also factor out some common code from the two clauses for different types of packets, and some 0/1 -> FALSE/TRUE cleanup.
in net.c, enforce that no packet has both NET_RELIABLE and
NET_DROPPABLE, both when packets are buffered, and also when they're
being sent.

this should finally fix the infamous lag bug, a.k.a. lag coma, a.k.a.
reliable sync bug.

the problem was that packets could be added to the outgoing queues that
were both reliable and droppable. if a player hit their bandwidth limit
at just the right time, one of these packets could be dropped without
ever being sent, creating a hole in the reliable stream, and the client
would be unable to process any reliable packets after the hole.